requirements.txt
selenium~=4.22.0
pandas~=2.2.3
python-dotenv~=1.0.1
requests~=2.32.5
schedule~=1.2.2
pillow~=12.1.1
PyMuPDF~=1.27.2
SQLAlchemy~=2.0.48
pyTelegramBotAPI~=4.32.0
numpy~=2.4.3
openai~=2.26.0
scikit-learn~=1.8.0
pymupdf-fonts~=1.0.5
boto3~=1.42.67
pre-commit~=4.5.1
cachetools~=7.0.5
botocore~=1.42.67
pytest~=9.0.2
pytest-cov~=7.0
coverage~=7.13.5
uvicorn~=0.41.0
fastapi~=0.135.1
pytest-xdist~=3.8.0